Lazy load the form descriptor
