+++ apps/application-system/form/src/routes/Application.tsx
@@ -1,9 +1,6 @@
 import React from 'react'
 import { useParams } from 'react-router-dom'
-
 import { ApplicationForm } from '@island.is/application/form'
-import { GET_APPLICATION } from '@island.is/application/graphql'
-import { useQuery } from '@apollo/client'
 import { useNamespaces } from '@island.is/localization'
 import useAuth from '../hooks/useAuth'
 
@@ -12,28 +9,19 @@ export const Application = () => {
   const { userInfo } = useAuth()
   useNamespaces(['dl.application', 'pl.application', 'application.system'])
 
-  const { data, error, loading } = useQuery(GET_APPLICATION, {
-    variables: {
-      input: {
-        id: id,
-      },
-    },
-    skip: !id,
-  })
+  const nationalRegistryId = userInfo?.profile?.natreg
 
   if (!id) {
     return <p>Error there is no id</p>
   }
-  if (error) {
-    return <p>{error}</p>
-  }
-  if (loading) {
+  if (!nationalRegistryId) {
     return null
   }
+
   return (
     <ApplicationForm
-      application={data.getApplication}
-      nationalRegistryId={userInfo?.profile?.natreg}
+      applicationId={id}
+      nationalRegistryId={nationalRegistryId}
     />
   )
 }
