+++ libs/application/template/src/templates/ApplicationTemplate.spec.ts
@@ -39,7 +39,7 @@ const TestApplicationTemplate: ApplicationTemplate<
   ApplicationStateSchema<TestEvents>,
   TestEvents
 > = {
-  mapNationalRegistryIdToRole(): ApplicationRole {
+  mapUserToRole(): ApplicationRole {
     return 'applicant'
   },
   type: ApplicationTypes.EXAMPLE,
@@ -64,7 +64,7 @@ const TestApplicationTemplate: ApplicationTemplate<
             {
               actions: [{ event: 'SUBMIT', name: 'Submit', type: 'primary' }],
               id: 'applicant',
-              form: ParentalLeaveForm,
+              formLoader: () => Promise.resolve(ParentalLeaveForm),
               write: {
                 answers: ['person', 'wantsInsurance'],
                 externalData: ['salary'],
