+++ libs/application/template/src/templates/ParentalLeave/index.ts
@@ -6,7 +6,6 @@ import {
   ApplicationStateSchema,
 } from '../../types/StateMachine'
 import * as z from 'zod'
-import { ParentalLeaveForm } from './ParentalLeaveForm'
 
 type Events =
   | { type: 'APPROVE' }
@@ -48,7 +47,10 @@ export const ParentalLeave: ApplicationTemplate<
           roles: [
             {
               id: 'applicant',
-              form: ParentalLeaveForm,
+              formLoader: () =>
+                import('./ParentalLeaveForm').then((val) =>
+                  Promise.resolve(val.ParentalLeaveForm),
+                ),
               actions: [{ event: 'SUBMIT', name: 'Submit', type: 'primary' }],
               write: {
                 answers: ['usage', 'spread', 'periods'],
@@ -109,7 +111,7 @@ export const ParentalLeave: ApplicationTemplate<
       },
     },
   },
-  mapNationalRegistryIdToRole(): ApplicationRole {
+  mapUserToRole(): ApplicationRole {
     return 'applicant'
   },
 }
